requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
databases
cryptography
